a bunch of fixes.

+++ client/app/following/following.service.js
@@ -6,27 +6,30 @@
       this.http = $http;
       this.auth = Auth;
     }
-
     getFollowers(campaign) {
-      this.http.get(`/api/campaigns/${campaign}/followers`).success(result => result.data);
+      return this.http.get(`/api/campaigns/${campaign}/followers`).success(result => result.data);
     }
-    getFollowings() {
-      this.http.get(`/api/my/followings`).success(result => result.data);
+    getMyFollowings() {
+      if (this.auth.isLoggedIn()) {
+        return this.http.get(`/api/my/followings`).success(result => result.data);
+      }
     }
     checkIfFollower(campaign) {
-      return !!this.getFollowings().then(data => { return _.filter(data, val => { return val === campaign; }) });
+      return this.getMyFollowings().then(data => { console.log(_.filter(data, val => {return val.campaign_id === campaign })) });
     }
-    follow() {
+    follow(campaign) {
       var _this = this;
-      var data = { 'user_id': _this.auth.getCurrentUser()._id, 'campaign_id': _this.campaign._id };
-      this.http.post('/api/followers', data).success(() => { _this.follow = true; });
+      if (this.auth.isLoggedIn()) {
+        var data = { 'user_id': _this.auth.getCurrentUser()._id, 'campaign_id': campaign };
+        return this.http.post('/api/followers', data);
+      }
     }
     unfollow(followid) {
-      this.http.delete(`/api/followers/${followid}`);
+      if (this.auth.isLoggedIn()) {
+        return this.http.delete(`/api/followers/${followid}`);
+      }
     }
   }
-
-
   angular.module('bApp')
     .service('followingFactory', followingFactory);
 })();
